[JBWS-3905]:Fix NPE;Refactor work to better name the new added interceptor
